Release V4R at version 2.0.8

+++ modules/io/include/v4r/io/eigen.h
@@ -13,7 +13,7 @@
 namespace v4r {
 namespace io {
 
-V4R_EXPORTS Eigen::Matrix4f readMatrixFromFile(const boost::filesystem::path &file, int padding = 0);
+V4R_EXPORTS Eigen::Matrix4f readMatrixFromFile(const boost::filesystem::path &file, size_t padding = 0);
 V4R_EXPORTS bool writeMatrixToFile(const boost::filesystem::path &, const Eigen::Matrix4f &);
-}
-}
+}  // namespace io
+}  // namespace v4r
